# prototipo-mqtt-multimod-3mods/api/src/main/resources/openapi/SwaggerIoTEste.yml
openapi: 3.0.4
info:
  title: API Sensores IoTEste
  description: >
    El sistema IoTEste EcoWarm busca mejorar el confort térmico y optimizar el consumo eléctrico en hogares y oficinas con calefacción por losa radiante.  
    La aplicación gestiona habitaciones que contienen sensores (que miden temperatura y humedad) y switches de calefacción (que miden consumo eléctrico).  
    Nuestra API permite crear habitaciones, consultar lecturas de sensores y consumos energéticos de switches.
  version: 1.3.2

servers:
  - url: http://RELLENARCONLAIPREAL

security:
  - basicAuth: []

paths:
  /rooms:
    post:
      summary: Crear una nueva habitación
      description: >
        Crea una nueva habitación indicando el número de sensores en el cuerpo del request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sensors:
                  type: integer
                  minimum: 1
                  description: Número de sensores a instalar en la habitación
                  example: 3
      responses:
        '201':
          description: Habitación creada con sus sensores asociados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
              example:
                id: "room-5"
                sensores:
                  - id: "1"
                  - id: "2"
                  - id: "3"
                switches:
                  - id: "1"

  /rooms/{roomId}/sensors/{sensorId}/temperature/latest:
    get:
      summary: Consultar temperatura (último registro)
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: string
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Última lectura de temperatura
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LecturaTemperatura'

  /rooms/{roomId}/sensors/{sensorId}/humidity/latest:
    get:
      summary: Consultar humedad (último registro)
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: string
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Última lectura de humedad
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LecturaHumedad'

  /rooms/{roomId}/sensors/{sensorId}/temperature/sequence:
    get:
      summary: Consultar secuencia de valores de temperatura
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: string
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
        - in: query
          name: fechaInicio
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: fechaFin
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Secuencia de lecturas de temperatura
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecuenciaTemperatura'

  /rooms/{roomId}/sensors/{sensorId}/humidity/sequence:
    get:
      summary: Consultar secuencia de valores de humedad
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: string
        - in: path
          name: sensorId
          required: true
          schema:
            type: string
        - in: query
          name: fechaInicio
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: fechaFin
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Secuencia de lecturas de humedad
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecuenciaHumedad'

  /rooms/{roomId}/switches/{switchId}/energy/consumption:
    get:
      summary: Consultar consumo energético (último día)
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: string
        - in: path
          name: switchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Consumo energético del último día
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnergiaShelly'

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  schemas:
    Room:
      type: object
      description: Representa una habitación con sensores y switches
      properties:
        id:
          type: string
          example: "room-1"
        sensores:
          type: array
          items:
            $ref: '#/components/schemas/Sensor'
        switches:
          type: array
          items:
            $ref: '#/components/schemas/Switch'

    Sensor:
      type: object
      description: Representa un sensor que mide temperatura y humedad
      properties:
        id:
          type: string
          example: "1"

    Switch:
      type: object
      description: Representa un switch de calefacción con medición de energía
      properties:
        id:
          type: string
          example: "1"

    LecturaTemperatura:
      type: object
      description: Lectura individual de temperatura
      properties:
        value:
          type: number
          format: float
          example: 22.4
        unit:
          type: string
          example: "°C"
        ts:
          type: string
          format: date-time
          example: "2025-09-23T15:30:00Z"

    LecturaHumedad:
      type: object
      description: Lectura individual de humedad
      properties:
        value:
          type: number
          format: float
          example: 45.2
        unit:
          type: string
          example: "%"
        ts:
          type: string
          format: date-time
          example: "2025-09-23T15:30:00Z"

    SecuenciaTemperatura:
      type: array
      description: Lista de lecturas de temperatura entre dos fechas
      items:
        $ref: '#/components/schemas/LecturaTemperatura'

    SecuenciaHumedad:
      type: array
      description: Lista de lecturas de humedad entre dos fechas
      items:
        $ref: '#/components/schemas/LecturaHumedad'

    EnergiaShelly:
      type: object
      description: Consumo energético del último día
      properties:
        value:
          type: number
          format: float
          example: 3.75
        unit:
          type: string
          example: "kWh"
        ts:
          type: string
          format: date-time
          example: "2025-09-23T00:00:00Z"